requirements.txt
asgiref==3.8.1
Django==5.2.1
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.2.0
django-import-export==2.0
django-filter==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
django-allauth==65.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
